+++ docs/changelog.md
@@ -1,7 +1,12 @@
 # TBD
 
+### Changes
+* Added `kurtosis-engine-server` to the monorepo
+* Merged the `kurtosis-engine-sdk` & `kurtosis-core-sdk`
+
 ### Removals
 * Remove unused variables from Docker Kurtosis backend
+
 
 # 0.49.4